generate_soilmap_simple: register non-spatial table *
Non-spatial tables need to be registered in the GeoPackage
in order to conform to the GPKG standard.
For now, this is done in a manual fashion, despite the ability
of GDAL to take care of this for GPKG. I found no way to use
this ability from R, so implemented the easy, though obsoleted
approach of GeoPackage 1.0 to register non-spatial tables as
data_type "aspatial" in the gpkg_contents table; see:
https://gdal.org/drivers/vector/aspatial.html#vector-aspatial

Question on implementation in R, see:
https://github.com/r-spatial/sf/issues/1345

The current way (GPKG 1.2) to handle this (data_type "attributes")
is supported by GDAL, see:
https://gdal.org/drivers/vector/gpkg.html#layer-creation-options .
But it needs more work to do it manually, so
made a GPKG 1.0 compliant version for now.
See https://www.ogc.org/standards/geopackage .

Note that GDAL is supportive when reading non-compliant
gpkg-files, as it 'will also, by default, list non spatial tables
that are not registered', which is why this worked beforehand
with st_read().
See https://gdal.org/drivers/vector/gpkg.html#non-spatial-tables .