generate_habitatquarries: drop variables 'type' and 'is_habitat' *
They overcomplicates the data source: 'code_orig', 'type' and
'is_habitat' are more appropriate to have as standardized, derived
variables in the resulting output of a function in n2khab.

'code_orig' has been renamed as 'habitattype', in accordance with
the approach followed for the raw data source 'habitatsprings'.

Such variables are more convenient to have in 'processed' data
sources than in a 'raw' data source.