shallowgw_update: alternative approach (sg_extended_2.gpkg) *
Compared to sg_extended.gpkg, several measures were taken to reduce the
influence of obliggwdep types on the result, notably by using a 50 m
instead of a 200 m buffer, and not using the obliggwdep polygons in the
ring. Instead of the latter, a 50 m buffer has been added and retained around
the original shallowgroundwater. So apart from that, only the (obliggwdep
bufferpolygons in) anthropogenic soil type polygons were added, with a buffer
of 50 m.