shallowgw_update: use new shallowgroundwater_20211129 as input *
Following previous naming, the output_filename would have been
sg3_extended_4.gpkg.
However we now consider '_extended' to be referring the latest
approach of extending, and we use the suffix '_20211129' to refer
the input shallowgroundwater layer.

Further, a parameter was added to select the appropriate layer
from the multilayer input GPKG.

Finally, the datestamp to get a reproducible GPKG file has been updated.